refactor(gateway): split admin handler monolith into cohesive modules (GW-1)
`src/handlers/admin.rs` was a 2874-LOC monolith mixing 5+ responsibilities.
Split it into a 80-LOC façade + 17 submodules under `src/handlers/admin/`,
each scoped to one concern (auth, health, apis, policies, cache,
circuit_breaker, contracts, credentials, federation, llm, mtls, quotas,
reload, sessions, skills, tracing).

- Tests inline preserved per-module (convention Rust)
- Shared test helpers centralised in `admin/test_helpers.rs`
- Contracts tests (16, ~466 LOC) split via `contracts.rs` + `contracts/tests.rs`
- Re-exports from `admin.rs` keep `crate::handlers::admin::<symbol>` stable;
  `lib.rs`, `main.rs`, `state.rs` unchanged
- `.gitignore`: anchor `/contracts/` + `/invoices/` to repo root so the
  legal-doc rule no longer swallows Rust submodules named `contracts/`

Validation: `cargo check`, `cargo fmt --check`, `cargo clippy --all-targets
-- -D warnings` all clean; full `cargo test` = 2336 passed / 0 failed;
`cargo test --lib handlers::admin` = 57 passed / 57 baseline (distribution
per submodule documented in `stoa-gateway/REWRITE-PLAN.md`).

0 behavioural change: routes, status codes, response schemas, auth
middleware, and public API surface unchanged.
