+++ .gitignore
@@ -126,8 +126,8 @@ emails/
 communication-*.md
 
 # Legal documents
-contracts/
-invoices/
+/contracts/
+/invoices/
 NDA-*.pdf
 MOU-*.pdf
 *-contract.pdf
