+++ stoa-gateway/src/handlers/admin/auth.rs
@@ -1,0 +1,151 @@
+//! Bearer-token admin authentication middleware.
+
+use axum::{
+    body::Body,
+    extract::State,
+    http::{header::AUTHORIZATION, Request, StatusCode},
+    middleware::Next,
+    response::{IntoResponse, Response},
+};
+use tracing::warn;
+
+use crate::state::AppState;
+
+/// Bearer token authentication for admin API.
+///
+/// Validates the `Authorization: Bearer <token>` header against
+/// `config.admin_api_token`. If no token is configured, returns 503
+/// (admin API disabled -- no token configured).
+pub async fn admin_auth(
+    State(state): State<AppState>,
+    request: Request<Body>,
+    next: Next,
+) -> Result<Response, Response> {
+    let expected = match state.config.admin_api_token.as_deref() {
+        Some(token) if !token.is_empty() => token,
+        _ => {
+            warn!("Admin API request rejected: no admin_api_token configured");
+            return Err((
+                StatusCode::SERVICE_UNAVAILABLE,
+                "Admin API disabled - no admin_api_token configured",
+            )
+                .into_response());
+        }
+    };
+
+    let auth_header = request
+        .headers()
+        .get(AUTHORIZATION)
+        .and_then(|v| v.to_str().ok())
+        .unwrap_or("");
+
+    let expected_header = format!("Bearer {}", expected);
+    if auth_header == expected_header {
+        Ok(next.run(request).await)
+    } else {
+        warn!("Admin API request rejected: invalid bearer token");
+        Err(StatusCode::UNAUTHORIZED.into_response())
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use axum::{
+        body::Body,
+        http::{Request, StatusCode},
+    };
+    use tower::ServiceExt;
+
+    use crate::handlers::admin::test_helpers::{build_admin_router, create_test_state};
+
+    #[tokio::test]
+    async fn test_admin_auth_valid_token() {
+        let state = create_test_state(Some("test-secret"));
+        let app = build_admin_router(state);
+
+        let response = app
+            .oneshot(
+                Request::builder()
+                    .uri("/health")
+                    .header("Authorization", "Bearer test-secret")
+                    .body(Body::empty())
+                    .unwrap(),
+            )
+            .await
+            .unwrap();
+
+        assert_eq!(response.status(), StatusCode::OK);
+    }
+
+    #[tokio::test]
+    async fn test_admin_auth_invalid_token() {
+        let state = create_test_state(Some("test-secret"));
+        let app = build_admin_router(state);
+
+        let response = app
+            .oneshot(
+                Request::builder()
+                    .uri("/health")
+                    .header("Authorization", "Bearer wrong-token")
+                    .body(Body::empty())
+                    .unwrap(),
+            )
+            .await
+            .unwrap();
+
+        assert_eq!(response.status(), StatusCode::UNAUTHORIZED);
+    }
+
+    #[tokio::test]
+    async fn test_admin_auth_no_token_configured() {
+        let state = create_test_state(None);
+        let app = build_admin_router(state);
+
+        let response = app
+            .oneshot(
+                Request::builder()
+                    .uri("/health")
+                    .header("Authorization", "Bearer anything")
+                    .body(Body::empty())
+                    .unwrap(),
+            )
+            .await
+            .unwrap();
+
+        assert_eq!(response.status(), StatusCode::SERVICE_UNAVAILABLE);
+    }
+
+    #[tokio::test]
+    async fn test_admin_auth_missing_header() {
+        let state = create_test_state(Some("secret"));
+        let app = build_admin_router(state);
+        let response = app
+            .oneshot(
+                Request::builder()
+                    .uri("/health")
+                    .body(Body::empty())
+                    .unwrap(),
+            )
+            .await
+            .unwrap();
+        assert_eq!(response.status(), StatusCode::UNAUTHORIZED);
+    }
+
+    #[tokio::test]
+    async fn test_admin_auth_empty_configured_token() {
+        let state = create_test_state(Some(""));
+        let app = build_admin_router(state);
+        let response = app
+            .oneshot(
+                Request::builder()
+                    .uri("/health")
+                    .header("Authorization", "Bearer ")
+                    .body(Body::empty())
+                    .unwrap(),
+            )
+            .await
+            .unwrap();
+        // Empty token = disabled
+        assert_eq!(response.status(), StatusCode::SERVICE_UNAVAILABLE);
+    }
+}
