+++ stoa-gateway/src/handlers/admin/health.rs
@@ -1,0 +1,95 @@
+//! Admin health endpoint.
+
+use axum::{extract::State, Json};
+use serde::Serialize;
+
+use crate::state::AppState;
+
+#[derive(Serialize)]
+pub struct AdminHealthResponse {
+    pub status: String,
+    pub version: String,
+    pub routes_count: usize,
+    pub policies_count: usize,
+    pub git_provider: String,
+}
+
+pub async fn admin_health(State(state): State<AppState>) -> Json<AdminHealthResponse> {
+    Json(AdminHealthResponse {
+        status: "ok".to_string(),
+        version: env!("CARGO_PKG_VERSION").to_string(),
+        routes_count: state.route_registry.count(),
+        policies_count: state.policy_registry.count(),
+        git_provider: state.config.git_provider.clone(),
+    })
+}
+
+#[cfg(test)]
+mod tests {
+    use axum::{
+        body::Body,
+        http::{Request, StatusCode},
+    };
+    use tower::ServiceExt;
+
+    use crate::config::Config;
+    use crate::handlers::admin::test_helpers::{build_admin_router, create_test_state};
+    use crate::state::AppState;
+
+    #[tokio::test]
+    async fn test_admin_health() {
+        let state = create_test_state(Some("secret"));
+        let app = build_admin_router(state);
+
+        let response = app
+            .oneshot(
+                Request::builder()
+                    .uri("/health")
+                    .header("Authorization", "Bearer secret")
+                    .body(Body::empty())
+                    .unwrap(),
+            )
+            .await
+            .unwrap();
+
+        assert_eq!(response.status(), StatusCode::OK);
+        let body = axum::body::to_bytes(response.into_body(), 1024 * 1024)
+            .await
+            .unwrap();
+        let data: serde_json::Value = serde_json::from_slice(&body).unwrap();
+        assert_eq!(data["status"], "ok");
+        assert!(data["version"].is_string());
+        assert_eq!(data["routes_count"], 0);
+        assert_eq!(data["policies_count"], 0);
+        assert_eq!(data["git_provider"], "gitlab");
+    }
+
+    #[tokio::test]
+    async fn test_admin_health_reports_github_provider() {
+        let config = Config {
+            admin_api_token: Some("secret".into()),
+            git_provider: "github".into(),
+            ..Config::default()
+        };
+        let state = AppState::new(config);
+        let app = build_admin_router(state);
+
+        let response = app
+            .oneshot(
+                Request::builder()
+                    .uri("/health")
+                    .header("Authorization", "Bearer secret")
+                    .body(Body::empty())
+                    .unwrap(),
+            )
+            .await
+            .unwrap();
+
+        assert_eq!(response.status(), StatusCode::OK);
+        let body = axum::body::to_bytes(response.into_body(), 1024 * 1024)
+            .await
+            .unwrap();
+        let data: serde_json::Value = serde_json::from_slice(&body).unwrap();
+        assert_eq!(data["git_provider"], "github");
+    }
+}
